Refactor handlers into mesh gateway handlers and router
